vote_app_backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-dotenv==1.4.2
django-cors-headers==3.11.0
mysqlclient==2.1.0
sqlparse==0.4.2
pytest-django==4.5.2
gunicorn==20.1.0